extract <player> component
